refactor: auto-split test_starting_pose_matcher.py
